requirements.txt
flask
redis
redisearch
python-dotenv
youtube-transcript-api
python-youtube
pyyaml
loguru